requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.61
attrs==25.3.0
bcrypt==4.1.2
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
confluent-kafka==2.10.0
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
dash==3.0.4
dash-bootstrap-components==2.0.2
dask==2025.4.1
distributed==2025.4.1
dnspython==2.7.0
entrypoints==0.4
exceptiongroup==1.2.2
Faker==37.1.0
fast-langdetect==0.3.2
fasttext-predict==0.9.2.4
favicon==0.7.0
feedparser==6.0.11
filelock==3.18.0
Flask==3.0.2
fonttools==4.57.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hf-xet==1.0.5
hpack==4.1.0
htbuilder==0.9.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kafka-python==2.1.5
kiwisolver==1.4.8
libipld==3.0.1
locket==1.0.0
lxml==5.4.0
Markdown==3.8
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.36.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.5
packaging==24.2
pandas==2.2.3
partd==1.4.2
pillow==11.2.1
plotly==6.0.1
portalocker==2.10.1
praw==7.8.1
prawcore==2.4.0
prometheus_client==0.21.1
protobuf==5.29.4
psutil==7.0.0
py4j==0.10.9.7
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
pymdown-extensions==10.15
pymongo==4.12.0
PyMySQL==1.1.0
pyparsing==3.2.3
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.14.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
robust-downloader==0.0.2
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
st-annotated-text==4.0.2
st-theme==1.2.3
streamlit==1.44.1
streamlit-autorefresh==1.0.1
streamlit-avatar==0.1.3
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.6.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
streamz==0.6.4
sympy==1.13.3
tblib==3.1.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
validators==0.35.0
watchdog==6.0.0
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.2
zict==3.0.0
zipp==3.21.0
